requirements.txt
chardet==5.2.0
coverage==5.5
deprecation==2.1.0
eyed3==0.9.7
filetype==1.2.0
packaging==24.2
pillow==11.1.0
Pmw==2.1.1
reportlab==4.3.1
toml==0.10.2
